requirements.txt
gradio
torch
transformers
sentencepiece
pdfplumber
python-docx

